winedbg: Fixed a few data types.

+++ programs/winedbg/be_i386.c
@@ -574,7 +574,7 @@ static unsigned be_i386_is_func_call(const void* insn, ADDRESS64* callee)
 #define	DR7_ENABLE_MASK(dr)	(1<<(DR7_LOCAL_ENABLE_SHIFT+DR7_ENABLE_SIZE*(dr)))
 #define	IS_DR7_SET(ctrl,dr) 	((ctrl)&DR7_ENABLE_MASK(dr))
 
-static inline int be_i386_get_unused_DR(CONTEXT* ctx, unsigned long** r)
+static inline int be_i386_get_unused_DR(CONTEXT* ctx, DWORD** r)
 {
     if (!IS_DR7_SET(ctx->Dr7, 0))
     {
@@ -607,7 +607,7 @@ static unsigned be_i386_insert_Xpoint(HANDLE hProcess, const struct be_process_i
 {
     unsigned char       ch;
     SIZE_T              sz;
-    unsigned long*      pr;
+    DWORD              *pr;
     int                 reg;
     unsigned long       bits;
 
@@ -630,7 +630,7 @@ static unsigned be_i386_insert_Xpoint(HANDLE hProcess, const struct be_process_i
         bits = DR7_RW_WRITE;
     hw_bp:
         if ((reg = be_i386_get_unused_DR(ctx, &pr)) == -1) return 0;
-        *pr = (unsigned long)addr;
+        *pr = (DWORD)addr;
         if (type != be_xpoint_watch_exec) switch (size)
         {
         case 4: bits |= DR7_LEN_4; break;
